API/python/requirements.txt
pyOSC
Flask
simplejson
